Introduce OS modules detection outside CrayPE (#10)
The default behaviour on new CSCS systems is to have no loaded modules
at all. If a user needs CrayPE, they need to load `cray` module first, a
CSCS provided module that takes care of "restoring" basic modules
available on a Cray system.

Part of this OS modules are also packages that are not part of CPE, e.g.
`libfabric`, which now can be detected as external spack packages as
well.